renamed test.js file
